Split cli_test into query and tx_test
